Add initial csv repository.
* Objective

This commit objective is about adding the repository needed to read the
Binance exported transactions CSV files.

* Why

As a starting point, we need a list of the Transactions for later
processing.

* How

This commit initializes the commands using Cobra and introduces a CSV
repository that will read CSV files for later processing.